requirements.txt
fastapi
uvicorn
python-dotenv
litellm
qdrant-client
sqlalchemy
psycopg2-binary
requests
pydantic
python-telegram-bot
neonize
rich
typer
httpx
apscheduler
questionary
markitdown
elevenlabs
